refactor(core): the "int backlight_level" member of struct "power_save_wakeup_params_t" has been refactored  to "uin8_t data type" in order to ensure consistency throughout the whole codebase.
[no changelog]

+++ core/embed/sys/suspend/inc/sys/suspend_io.h
@@ -42,7 +42,7 @@ void suspend_cpu(void);
 typedef struct {
 #ifdef USE_BACKLIGHT
   /** Backlight level */
-  int backlight_level;
+  uint8_t backlight_level;
 #endif
 #ifdef USE_BLE
   /** State of the ble driver */
